backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
argilla==1.7.0
async-timeout==4.0.2
attrs==23.1.0
backoff==2.2.1
black==23.3.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.1.0
chromadb==0.3.25
click==8.1.3
clickhouse-connect==0.5.25
coloredlogs==15.0.1
commonmark==0.9.1
cryptography==40.0.2
dataclasses-json==0.5.7
Deprecated==1.2.13
dnspython==2.3.0
duckdb==0.8.0
elastic-transport==8.4.0
elasticsearch==8.7.0
et-xmlfile==1.1.0
faiss-cpu==1.7.4
fastapi==0.95.2
flake8==6.0.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
flatbuffers==23.5.9
frozenlist==1.3.3
fsspec==2023.5.0
greenlet==2.0.2
h11==0.14.0
hnswlib==0.7.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.6.0
isort==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
langchain==0.0.185
llama-index==0.6.14
lxml==4.9.2
lz4==4.3.2
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mccabe==0.7.0
monotonic==1.6
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.8.4
numpy==1.23.5
olefile==0.46
onnxruntime==1.15.0
openai==0.27.7
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
overrides==7.3.1
packaging==23.1
pandas==1.5.3
pathspec==0.11.1
pdf2image==1.16.3
pdfminer.six==20221105
Pillow==9.5.0
platformdirs==3.5.1
posthog==3.0.1
protobuf==4.23.1
pycodestyle==2.10.0
pycparser==2.21
pydantic==1.10.8
pyflakes==3.0.1
Pygments==2.15.1
pymongo==4.3.3
pypandoc==1.11
pypdf==3.9.0
pypdf2==3.0.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-magic==0.4.27
python-pptx==0.6.21
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
rfc3986==1.5.0
rich==13.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
tiktoken==0.4.0
tokenizers==0.13.3
tomli==2.0.1
tqdm==4.65.0
typer==0.9.0
typing-extensions==4.5.0
typing-inspect==0.8.0
tzdata==2023.3
unstructured==0.6.9
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
Werkzeug==2.3.4
wrapt==1.14.1
XlsxWriter==3.1.1
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
